+++ assets/js/pages/services.js
@@ -98,6 +98,8 @@ App.Pages.Services = (function () {
             $filterServices.find('button').prop('disabled', true);
             $filterServices.find('.results').css('color', '#AAA');
             App.Components.ColorSelection.enable($color);
+            $('#service-providers input:checkbox').prop('disabled', false);
+            $('#select-all-providers, #select-none-providers').prop('disabled', false);
         });
 
         /**
@@ -114,6 +116,8 @@ App.Pages.Services = (function () {
             $filterServices.find('button').prop('disabled', true);
             $filterServices.find('.results').css('color', '#AAA');
             App.Components.ColorSelection.enable($color);
+            $('#service-providers input:checkbox').prop('disabled', false);
+            $('#select-all-providers, #select-none-providers').prop('disabled', false);
 
             // Default values
             $name.val('Service');
@@ -159,6 +163,14 @@ App.Pages.Services = (function () {
                 id_service_categories: $serviceCategoryId.val() || undefined,
             };
 
+            // Include service providers.
+            service.providers = [];
+            $('#service-providers input:checkbox').each((index, checkboxEl) => {
+                if ($(checkboxEl).prop('checked')) {
+                    service.providers.push($(checkboxEl).attr('data-id'));
+                }
+            });
+
             if ($id.val() !== '') {
                 service.id = $id.val();
             }
@@ -182,6 +194,8 @@ App.Pages.Services = (function () {
             $filterServices.find('button').prop('disabled', true);
             $filterServices.find('.results').css('color', '#AAA');
             App.Components.ColorSelection.enable($color);
+            $('#service-providers input:checkbox').prop('disabled', false);
+            $('#select-all-providers, #select-none-providers').prop('disabled', false);
         });
 
         /**
@@ -206,6 +220,20 @@ App.Pages.Services = (function () {
             ];
 
             App.Utils.Message.show(lang('delete_service'), lang('delete_record_prompt'), buttons);
+        });
+
+        /**
+         * Event: Select All Providers Button "Click"
+         */
+        $services.on('click', '#select-all-providers', () => {
+            $('#service-providers input:checkbox').prop('checked', true);
+        });
+
+        /**
+         * Event: Select None Providers Button "Click"
+         */
+        $services.on('click', '#select-none-providers', () => {
+            $('#service-providers input:checkbox').prop('checked', false);
         });
     }
 
@@ -296,6 +324,11 @@ App.Pages.Services = (function () {
         $services.find('.record-details .is-invalid').removeClass('is-invalid');
         $services.find('.record-details .form-message').hide();
 
+        // Reset providers checkboxes
+        $('#service-providers input:checkbox').prop('disabled', true).prop('checked', false);
+        $('#select-all-providers, #select-none-providers').prop('disabled', true);
+        $('#service-providers a').remove();
+
         App.Components.ColorSelection.disable($color);
     }
 
@@ -319,6 +352,43 @@ App.Pages.Services = (function () {
 
         const serviceCategoryId = service.id_service_categories !== null ? service.id_service_categories : '';
         $serviceCategoryId.val(serviceCategoryId);
+
+        // Display providers
+        $('#service-providers a').remove();
+        $('#service-providers input:checkbox').prop('checked', false);
+
+        if (service.providers) {
+            service.providers.forEach((serviceProviderId) => {
+                const $checkbox = $('#service-providers input[data-id="' + serviceProviderId + '"]');
+
+                if (!$checkbox.length) {
+                    return;
+                }
+
+                $checkbox.prop('checked', true);
+
+                // Add dedicated service-provider link.
+                const dedicatedUrl = App.Utils.Url.siteUrl(
+                    '?service=' + encodeURIComponent(service.id) + '&provider=' + encodeURIComponent(serviceProviderId),
+                );
+
+                const $link = $('<a/>', {
+                    'href': dedicatedUrl,
+                    'target': '_blank',
+                    'html': [
+                        $('<i/>', {
+                            'class': 'fas fa-link me-2',
+                        }),
+
+                        $('<span/>', {
+                            'text': lang('booking_link'),
+                        }),
+                    ],
+                });
+
+                $checkbox.parent().append($link);
+            });
+        }
     }
 
     /**